fix(grpc/snapshot): snapshot grpc should use replica svc
We should not allow concurrent snapshot and replica operations as they might clash.
todo: allow per-resource locking
Furthermore it was observed that holding a bdev ptr whilst it's being deleted might
cause a crash, we need to determine if this was to do with it being held over an
async point, or a more general problem.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>