fix(pstor): increase persistence timeouts
The pstor might might unavailable for some time, for example during upgrade.
We should have sufficient timeouts to cope with this.
Also pstor can get slow at times as it's writing to disk, so we also should
have a larger timeout on store.

todo: should we still carry on trying to persist on separate task after
we fail the nexus?
todo: what happens when we try to shutdown nexus first and pstor fails?

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>