refactor(rebuild): make the rebuild more generic
A rebuild is essentially a copy from one bdev to another.
However, we can have different variations on this.
This change aims to make the rebuild more generic, removing the nexus specific
bits from the core rebuild logic and allowing us to compose different types of
rebuild in the future, whilst being able to reuse the shared bits.

To achieve this the rebuild backend is split into a rebuild manager which is the
generic component responsible for running and managing the rebuild and its tasks.
We can then implement different rebuild backends as we see fit, example a nexus
rebuild which locks ranges or a regular bdev to bdev rebuild.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>