refactor(rebuild): use new rebuild rangers
Make use of the rebuild rangers to configure rebuild types.
This allows us to remove the setting of the rebuild map being done
after the rebuild job is created for the nexus and removing it from
the shared rebuild descriptor.

The nexus still uses the partial but sequential rebuild to reduce
the scope of changes.
Once the fully partial rebuild is validated we can switch the nexus
to it.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>