fix: update blob metadata during properties update
When a nexus fails over and is moved to a new node, the nvmf replicas attached
to it are re-shared. In the share path we don't update lvol blob metadata today,
but only update in-core replica nvmf subsystem. Later on if this replica node
reboots, the replica is imported with metadata that still contains nqn of old
nexus node. This results into replica rejecting the connect attempts from
nexus(on new node), e.g. during online child operation.

Signed-off-by: Diwakar Sharma <diwakar.sharma@datacore.com>