chore(wiper): temporarily reduce replica checksum size
This change is required because otherwise e2e checksum tests incorrectly
report data mismatch.
To better explain this, the nexus used to format replicas with a gpt
partition table. This is no longer the case, but for back-compat the nexus
ignores the blocks which were used for this.
As such the gpt backup is never read or written to by an application and
it is also not rebuilt! So this data may not be the same among the nexus
replicas.

As a short term solution, we simply discount these last 33 sectors from
the replica checksum.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>