fix(rebuild): reuse rebuild IO handles
Reuses the rebuild IO handles, rather than attempting to allocate
them per rebuild task.
The main issue with handle allocation on the fly is that the target
may have not cleaned up a previous IO qpair connection, and so the
connect may fail. We started seeing this more on CI because we forgot
to cherry-pick a commit increasing the retry delay.
However, after inspecting a bunch of user support bundles I see that
we still have occasional connect errors. Rather than increasing the
timeout, we attempt here to reuse the handles, thus avoid the
problem almost entirely.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>