fix(tgt-iface): support multiple ips per adapter
In case adapter has multiple ipv6 ips, then the libc function
getifaddrs returns multiple elements with the same name and so
putting them on a map was actually erasing some ips!

However, this now means we need to handle trying to pick one
of the ips!
For now we prioritize ipv4 vs 6 IF the grpc is also ipv4.
Also we prefer ipv6 unique local, link local and !multicast.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>