fix(tgt-iface): reject scope ids on ipv6
Start to add support for scope ids on ipv6.
This is not yet supported on the nvmf target, I suspect because there's
a bug on the socket interface which doesn't return the scope id as part
of the traddr, though this is mostly a guess so far.
As such, for the moment we detect but reject scope ids.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>