test: ignore test which is relying on absent behaviour
The test writes to the nexus bdev directly in a retire scenario where
frozen IOs are resubmitted.
The problem happens when we pause the nexus and drop the io channels
because pause only pauses the nvmf subsystem!!
So, if the IO which is resubmitted hasn't completed, then spdk will
assert because it expects IOs to be drained.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>