fix(bdev/destroy): unload pool when disk is hotremoved
When destroying a bdev as part of a hot-removal, the uri used to
destroy the bdev may not match exactly the alias which was added on
the creation.
In this case we'll leak a bdev alias which will cause issues if we
attempt to re-create the bdev as it'll have a stale alias pointer.

We've modified spdk to automatically delete all aliases on unregister,
and so we don't need to manually remove them, thus fixing this
scenario.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>